(pas fini) On mets a jour la scène et les objets

+++ Assets/Created/Scripts/SO/MillSO.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 public class MillSO : ItemSO
 {
     public int nbMillsByDefault;
-    public int costMultiplier;
+    public float costMultiplier;
 }
