+++ Assets/Created/Scripts/SO/SlotSO.cs
@@ -5,5 +5,5 @@ public class SlotSO : ItemSO
 {
     public int nbSlotMax;
     public int nbSlotByDefault;
-    public int costMultiplier;
+    public float costMultiplier;
 }
